+++ R/objective.R
@@ -89,6 +89,7 @@ Objective <- R6::R6Class(
     #' Invisibly the \code{Objective} object.
     set_argument = function(..., overwrite = TRUE, verbose = self$verbose) {
       checkmate::assert_flag(overwrite)
+      checkmate::assert_flag(verbose)
       arguments <- list(...)
       checkmate::check_names(arguments, type = "strict")
       argument_names <- names(arguments)
@@ -119,9 +120,10 @@ Objective <- R6::R6Class(
     #' Get a fixed function argument.
     #' @return
     #' The argument value.
-    get_argument = function(argument_name) {
+    get_argument = function(argument_name, verbose = self$verbose) {
       private$.check_argument_specified(argument_name)
-      if (self$verbose) {
+      checkmate::assert_flag(verbose)
+      if (verbose) {
         cli::cli_alert("Returning argument {.val {argument_name}}.")
       }
       private$.arguments[[argument_name]]
@@ -131,9 +133,10 @@ Objective <- R6::R6Class(
     #' Remove a fixed function argument.
     #' @return
     #' Invisibly the \code{Objective} object.
-    remove_argument = function(argument_name) {
+    remove_argument = function(argument_name, verbose = self$verbose) {
       private$.check_argument_specified(argument_name)
-      if (self$verbose) {
+      checkmate::assert_flag(verbose)
+      if (verbose) {
         cli::cli_alert("Removing argument {.val {argument_name}}.")
       }
       private$.arguments[[argument_name]] <- NULL